organize graphql lib
